requirements.txt
django-bootstrap3-datetimepicker
django== 3.0.8
django-bootstrap-datepicker-plus

django-bootstrap4
